[refactor]: 기존 Group 검색 컴포넌트 이름 및 divider 수정 (#47)
