+++ app/src/main/java/com/texthip/thip/ui/group/myroom/component/GroupLiveSearchResult.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.ui.group.myroom.screen
+package com.texthip.thip.ui.group.myroom.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -6,7 +6,7 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
@@ -15,13 +15,13 @@ import com.texthip.thip.ui.group.myroom.mock.GroupCardItemRoomData
 import com.texthip.thip.ui.theme.ThipTheme.colors
 
 @Composable
-fun GroupLiveSearchResultScreen(
+fun GroupLiveSearchResult(
     roomList: List<GroupCardItemRoomData>
 ) {
     LazyColumn(
         verticalArrangement = Arrangement.spacedBy(16.dp)
     ) {
-        items(roomList) { room ->
+        itemsIndexed(roomList) { index, room ->
             CardItemRoomSmall(
                 title = room.title,
                 participants = room.participants,
@@ -31,12 +31,14 @@ fun GroupLiveSearchResultScreen(
                 isWide = true,
                 isSecret = room.isSecret
             )
-            Spacer(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(1.dp)
-                    .background(colors.DarkGrey02)
-            )
+            if (index < roomList.size - 1) {
+                Spacer(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(1.dp)
+                        .background(colors.DarkGrey02)
+                )
+            }
         }
     }
 }
