+++ app/src/main/java/com/texthip/thip/ui/group/myroom/screen/GroupSearchScreen.kt
@@ -28,6 +28,10 @@ import com.texthip.thip.R
 import com.texthip.thip.ui.common.buttons.FilterButton
 import com.texthip.thip.ui.common.forms.SearchBookTextField
 import com.texthip.thip.ui.common.topappbar.DefaultTopAppBar
+import com.texthip.thip.ui.group.myroom.component.GroupEmptyResult
+import com.texthip.thip.ui.group.myroom.component.GroupFilteredSearchResult
+import com.texthip.thip.ui.group.myroom.component.GroupLiveSearchResult
+import com.texthip.thip.ui.group.myroom.component.GroupRecentSearch
 import com.texthip.thip.ui.group.myroom.mock.GroupCardItemRoomData
 import com.texthip.thip.ui.theme.ThipTheme
 
@@ -132,14 +136,14 @@ fun GroupSearchScreen(
 
                 when {
                     searchText.isBlank() && !isSearched && recentSearches.isEmpty() -> {
-                        GroupRecentSearchScreen(
+                        GroupRecentSearch(
                             recentSearches = emptyList(),
                             onSearchClick = {},
                             onRemove = {}
                         )
                     }
                     searchText.isBlank() && !isSearched && recentSearches.isNotEmpty() -> {
-                        GroupRecentSearchScreen(
+                        GroupRecentSearch(
                             recentSearches = recentSearches,
                             onSearchClick = { keyword ->
                                 searchText = keyword
@@ -152,18 +156,18 @@ fun GroupSearchScreen(
                     }
                     searchText.isNotBlank() && !isSearched -> {
                         if (liveFilteredRoomList.isEmpty()) {
-                            GroupEmptyResultScreen(
+                            GroupEmptyResult(
                                 mainText = stringResource(R.string.group_no_search_result1),
                                 subText = stringResource(R.string.group_no_search_result2)
                             )
                         } else {
-                            GroupLiveSearchResultScreen(
+                            GroupLiveSearchResult(
                                 roomList = liveFilteredRoomList
                             )
                         }
                     }
                     isSearched -> {
-                        GroupFilteredSearchResultScreen(
+                        GroupFilteredSearchResult(
                             genres = genres,
                             selectedGenreIndex = selectedGenreIndex,
                             onGenreSelect = { selectedGenreIndex = it },
