+++ app/src/main/java/com/texthip/thip/ui/group/screen/GroupScreen.kt
@@ -1,14 +1,12 @@
 package com.texthip.thip.ui.group.screen
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
@@ -33,7 +31,6 @@ import com.texthip.thip.ui.group.myroom.component.GroupSearchTextField
 import com.texthip.thip.ui.group.myroom.mock.GroupViewModel
 import com.texthip.thip.ui.theme.ThipTheme
 import com.texthip.thip.ui.theme.ThipTheme.colors
-
 
 @Composable
 fun GroupScreen(
